Moved methods to /d3rlpy_methods
