+++ sequoia/settings/offline_rl/setting.py
@@ -18,14 +18,24 @@ from d3rlpy.metrics.scorer import average_value_estimation_scorer
 import gym
 from simple_parsing.helpers import choice
 import re
-# get_cartpole
 
-# get_pendulum
+
+@dataclass()
+class OfflineRLResults(Results):
+    objective_name: ClassVar[str] = "Average Reward"
+    rewards: List[int]
+
+    @property
+    def objective(self):
+        assert self.rewards
+        return sum(self.rewards)/len(self.rewards)
+
+
+
 
 
 @dataclass
 class OfflineRLSetting(Setting):
-
     # We can pass in any of the available datasets shown below.
     # get_dataset() will attempt to match the regex expressions for d4rl, py_bullet and atari
 
@@ -33,14 +43,13 @@ class OfflineRLSetting(Setting):
         https://github.com/takuseno/d3rlpy/blob/master/d3rlpy/datasets.py
      """
 
+    # TODO: Results: ClassVar[Type[Results]] = OfflineRLResults
+
     available_datasets: ClassVar[list] = ["cartpole-replay",  # Cartpole Replay
                                           "cartpole-random",  # Cartpole Random
                                           "pendulum-replay",  # Pendulum Replay
                                           "pendulum-random",  # Pendulum Random
-                                          r"^bullet-.+$",     # d4rl
-                                          r"hopper|halfcheetah|walker|ant",  # also d4rl
-                                          r"^.+-bullet-.+$"  # py_bullet
-                                          # Atari: see d3rlpy.datasets.ATARI_GAMES
+
                                           ]
     dataset: str = choice(available_datasets, default="cartpole-replay")
     create_mask: bool = False
@@ -62,123 +71,10 @@ class OfflineRLSetting(Setting):
         method.fit(train_env=self.train_dataset,
                    valid_env=self.valid_dataset)
 
-
-class SequoiaToGymWrapper(gym.Wrapper):
-    def __init__(self, env):
-        super().__init__(env)
-        self.observation_space = env.observation_space.x
-
-        # TODO: If action space is changed to dictionary, do this
-        # self.action_space = env.action_space.y_pred
-
-    def reset(self):
-        observation = super().reset()
-        return observation.x
-
-    def step(self, action):
-        # TODO: if step expects a dictionary as action, just pass {'y_pred': action}
-        observation, reward, done, info = super().step(action)
-        return observation.x, reward.y, done, info
-
-
-class BaseOfflineRLMethod(Method, target_setting=OfflineRLSetting):
-
-    Algo: Type[AlgoBase] = AlgoBase
-
-    def __init__(self, train_steps: int = 1_000_000, n_epochs: int = 5, scorers: dict = None):
-        super().__init__()
-        self.train_steps = train_steps
-        self.n_epochs = n_epochs
-        self.scorers = scorers
-        self.algo = type(self).Algo()
-
-    def configure(self, setting: OfflineRLSetting):
-        super().configure(setting)
-        self.setting = setting
-
-    def fit(self, train_env, valid_env) -> None:
-        if isinstance(self.setting, OfflineRLSetting):
-            self.algo.fit(train_env,
-                          eval_episodes=valid_env,
-                          n_epochs=self.n_epochs,
-                          scorers=self.scorers)
-        else:
-            train_env, valid_env = SequoiaToGymWrapper(train_env), SequoiaToGymWrapper(valid_env)
-            self.algo.fit_online(env=train_env, eval_env=valid_env, n_steps=self.train_steps)
-
-    def get_actions(self, obs: np.ndarray, action_space: Space) -> np.ndarray:
-        # ready to control
-        return self.algo.predict(obs)
-
-
-"""
-D3RLPY Methods: work on OfflineRL and TraditionalRL assumptions
-"""
-
-
-class DQNMethod(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = DQN
-
-
-class DoubleDQNMethod(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = DoubleDQN
-
-
-class DDPGMethod(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = DDPG
-
-
-class TD3Method(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = TD3
-
-
-class SACMethod(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = SAC
-
-
-class DiscreteSACMethod(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = DiscreteSAC
-
-
-class CQLMethod(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = CQL
-
-
-class DiscreteCQLMethod(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = DiscreteCQL
-
-
-class BEAR(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = BEAR
-
-
-class AWRMethod(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = AWR
-
-
-class DiscreteAWRMethod(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = DiscreteAWR
-
-
-class BC(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = BC
-
-
-class DiscreteBCMethod(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = DiscreteBC
-
-
-class BCQMethod(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = BCQ
-
-
-class DiscreteBCQMethod(BaseOfflineRLMethod):
-    Algo: Type[AlgoBase] = DiscreteBCQ
-
-
 """
 Quick example using DQN for offline cart-pole
 """
+
 
 def main():
     setting_offline = OfflineRLSetting(dataset="cartpole-replay")
@@ -189,8 +85,8 @@ def main():
     })
 
     _ = setting_offline.apply(method)
-    results = setting_online.apply(method)
-    print(results)
+    # results = setting_online.apply(method)
+    # print(results)
 
 
 if __name__ == "__main__":
